Move tests to better location
